% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.bootstrapped_models}
\alias{predict.bootstrapped_models}
\title{Calculate bootstrap prediciton intervals for new samples}
\usage{
\method{predict}{bootstrapped_models}(x, new_data, ..., interval_width)
}
\description{
Calculate bootstrap prediciton intervals for new samples
}
