% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_boots.R
\name{predict_single_boot}
\alias{predict_single_boot}
\title{Fit a model and predict based on a single bootstrap resample}
\usage{
predict_single_boot(workflow, boot_splits, new_data, index)
}
\arguments{
\item{workflow}{An un-fitted workflow object.}

\item{boot_splits}{A bootstrap split object created by \code{rsample::bootstraps()}.}

\item{new_data}{New data to make predictions}

\item{index}{Index of \code{boot_splits} to use for training}
}
\description{
Fit a model and predict based on a single bootstrap resample
}
