% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_boots.R
\name{summarise_generic}
\alias{summarise_generic}
\title{(Internal) Function for generating generic summaries from either \code{predict_boots()}
or \code{vi_boots()}.}
\usage{
summarise_generic(.data, nest_col, conf)
}
\arguments{
\item{.data}{passed from one of the summarise_* functions}

\item{nest_col}{passed from one of the summarise_* functions}

\item{conf}{passed from one of the summarise_* functions}
}
\description{
(Internal) Function for generating generic summaries from either \code{predict_boots()}
or \code{vi_boots()}.
}
