% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_boots.R
\name{vi_single_boot}
\alias{vi_single_boot}
\title{Fit a model and get the variable importance based on a single bootstrap resample}
\usage{
vi_single_boot(workflow, boot_splits, verbose, index)
}
\arguments{
\item{workflow}{passed from \code{vi_boots()}}

\item{boot_splits}{passed from \code{vi_boots()}}

\item{verbose}{passed from \code{vi_boots()}}

\item{index}{passed from \code{vi_boots()}}
}
\description{
Fit a model and get the variable importance based on a single bootstrap resample
}
