% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bootstrap_models.R
\name{add_bootstrap_models}
\alias{add_bootstrap_models}
\title{Fit from a workflow using many bootstrap resamples.}
\usage{
add_bootstrap_models(
  resamples,
  workflow,
  interval = c("prediction", "confidence"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{resamples}{A resampling object created by \code{\link[rsample:bootstraps]{rsample::bootstraps()}}. If
the option \code{apparent = TRUE} was used to create it, the corresponding row
is removed.}

\item{workflow}{An un-fitted workflow object.}

\item{interval}{One of \code{prediction}, \code{confidence}. Specifies the interval
type to be generated.}

\item{verbose}{A logical. Defaults to \code{FALSE}. If set to \code{TRUE}, prints progress
of training to console.}

\item{...}{Additional params passed to \code{rsample::bootstraps()}.}
}
\description{
Generate a prediction interval from arbitrary model types using bootstrap
resampling. \code{add_bootstrap_models()} fits a model to each resample.
}
