% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_boots.R
\name{predict_single_boot}
\alias{predict_single_boot}
\title{(Internal) Generate a column of predictions on new data based on a model fit
to a single training bootstrap.}
\usage{
predict_single_boot(
  workflow,
  boot_splits,
  new_data,
  normal_resid,
  verbose,
  index
)
}
\arguments{
\item{workflow}{passed from \code{predict_boots()}}

\item{boot_splits}{passed from \code{predict_boots()}}

\item{new_data}{passed from \code{predict_boots()}}

\item{normal_resid}{passed from \code{predict_boots()}}

\item{verbose}{passed from \code{predict_boots()}}

\item{index}{passed from \code{predict_boots()}}
}
\description{
(Internal) Generate a column of predictions on new data based on a model fit
to a single training bootstrap.
}
