% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_boots.R
\name{vi_single_boot}
\alias{vi_single_boot}
\title{Fit a model and get the variable importance based on a single bootstrap resample}
\usage{
vi_single_boot(workflow, boot_splits, index)
}
\arguments{
\item{workflow}{An un-fitted workflow object.}

\item{boot_splits}{A bootstrap split object created by \code{rsample::bootstraps()}.}

\item{index}{Index of \code{boot_splits} to use for training}
}
\description{
Fit a model and get the variable importance based on a single bootstrap resample
}
